# Conflicts:
#	src/Simulation/MolflowSimGeom.cpp
#	src_shared

+++ src/FlowCLI.cpp
@@ -1,6 +1,22 @@
-//
-// Created by Pascal Baehr on 28.04.20.
-//
+/*
+Program:     MolFlow+ / Synrad+
+Description: Monte Carlo simulator for ultra-high vacuum and synchrotron radiation
+Authors:     Jean-Luc PONS / Roberto KERSEVAN / Marton ADY / Pascal BAEHR
+Copyright:   E.S.R.F / CERN
+Website:     https://cern.ch/molflow
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation; either version 2 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+Full license text: https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html
+*/
 
 // M_PI define
 #if defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
@@ -16,7 +32,7 @@
 #include <Parameter.h>
 #include <IO/LoaderXML.h>
 #include <IO/WriterXML.h>
-#include "GeometrySimu.h"
+#include "Simulation/MolflowSimGeom.h"
 #include "Initializer.h"
 #include "Helper/MathTools.h"
 #include <sstream>
@@ -40,17 +56,62 @@ static constexpr const char* molflowCliLogo = R"(
  |_|  |_\___/_|_| |_\___/\_/\_/
     )";
 
-void GatherResults(SimulationModel& model, GlobalSimuState& globSim){
+void GatherResults(MolflowSimulationModel& model, GlobalSimuState& globSim){
     for(int i = 0; i < model.facets.size(); i++ ) {
 #if defined(MOLFLOW)
-        auto &f = model.facets[i];
+        auto f = std::dynamic_pointer_cast<MolflowSimFacet>(model.facets[i]);
         if (f->sh.anglemapParams.record) { //Recording, so needs to be updated
             //Retrieve angle map from hits dp
-            model.facets[i]->angleMap.pdf = globSim.facetStates[i].recordedAngleMapPdf;
+            f->angleMap.pdf = globSim.facetStates[i].recordedAngleMapPdf;
         }
 #endif
     }
 }
+
+class RuntimeStatPrinter {
+    size_t oldHitsNb{0};
+    size_t oldDesNb{0};
+public:
+    RuntimeStatPrinter() = default;
+    RuntimeStatPrinter(size_t n_hits, size_t n_des) {
+        oldHitsNb = n_hits;
+        oldDesNb = n_des;
+    };
+    void PrintHeader() const{
+        // Print Header at the beginning
+        Log::console_msg_master(1, "\n");
+        if(MFMPI::world_size > 1) {
+            Log::console_msg_master(1, "%-6s ",
+                                    "Node#");
+        }
+        Log::console_msg_master(1, "%-14s %-20s %-20s %-20s %-20s %-20s %-20s\n",
+                                "Time",
+                                "#Hits (run)", "#Hits (total)","Hit/sec",
+                                "#Des (run)", "#Des (total)","Des/sec");
+        if(MFMPI::world_size > 1) {
+            Log::console_msg_master(1, "%s",std::string(6,'-').c_str());
+        }
+        Log::console_msg_master(1, "%s\n",std::string(14+20+20+20+20+20+20,'-').c_str());
+    }
+    void Print(double elapsedTime, GlobalSimuState& globState, bool printSum=false) const{
+        if(printSum) {
+            Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'=').c_str());
+            Log::console_msg_master(1, "%-6s ", "x");
+        }
+        else if(MFMPI::world_size > 1) {
+            Log::console_msg(1, "%-6d ", MFMPI::world_rank);
+        }
+
+        Log::console_msg(1,"%-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
+                         elapsedTime,
+                         globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
+                         (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
+                         (elapsedTime),
+                         globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
+                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
+                         (elapsedTime));
+    }
+};
 
 int main(int argc, char** argv) {
 
@@ -68,7 +129,7 @@ int main(int argc, char** argv) {
 
     SimulationManager simManager{MFMPI::world_rank};
     simManager.interactiveMode = true;
-    std::shared_ptr<SimulationModel> model = std::make_shared<SimulationModel>();
+    std::shared_ptr<MolflowSimulationModel> model = std::make_shared<MolflowSimulationModel>();
     GlobalSimuState globState{};
 
 
@@ -105,14 +166,17 @@ int main(int argc, char** argv) {
     }
     size_t oldHitsNb = globState.globalHits.globalHits.nbMCHit;
     size_t oldDesNb = globState.globalHits.globalHits.nbDesorbed;
-
+    RuntimeStatPrinter printer(oldHitsNb, oldDesNb);
     // Get autosave file name
     std::string autoSave = Initializer::getAutosaveFile();
 
 
     //simManager.ReloadHitBuffer();
     //simManager.IncreasePriority();
-    Log::console_msg_master(1,"[%s] Commencing simulation for %lu seconds from %lu desorptions.\n", Util::getTimepointString().c_str(), Settings::simDuration, globState.globalHits.globalHits.nbDesorbed);
+    if(Settings::simDuration > 0)
+        Log::console_msg_master(1,"[%s] Commencing simulation for %lu seconds from %lu desorptions.\n", Util::getTimepointString().c_str(), Settings::simDuration, globState.globalHits.globalHits.nbDesorbed);
+    else if(model->otfParams.desorptionLimit > 0)
+        Log::console_msg_master(1,"[%s] Commencing simulation to %lu desorptions from %lu desorptions.\n", Util::getTimepointString().c_str(), model->otfParams.desorptionLimit, globState.globalHits.globalHits.nbDesorbed);
 
 #if defined(USE_MPI)
     MPI_Barrier(MPI_COMM_WORLD);
@@ -144,40 +208,50 @@ int main(int argc, char** argv) {
             endCondition = globState.globalHits.globalHits.nbDesorbed/* - oldDesNb*/ >= model->otfParams.desorptionLimit;
 
         if(endCondition){
-            /*std::stringstream outFile;
-            outFile << SettingsIO::outputPath << "/" <<"desorped_" << model->otfParams.desorptionLimit << "_" <<
-                 std::filesystem::path(SettingsIO::outputFile).filename().string();*/
-            std::string outFile = std::filesystem::path(SettingsIO::outputPath)
-                    .append("desorped_")
-                    .concat(std::to_string(model->otfParams.desorptionLimit))
-                    .concat("_")
-                    .concat(std::filesystem::path(SettingsIO::outputFile).filename().string()).string();
-
-            try {
-                FlowIO::WriterXML writer;
-                if(!SettingsIO::workFile.empty()) {
-                    std::filesystem::copy_file(SettingsIO::workFile, outFile,
-                                               std::filesystem::copy_options::overwrite_existing);
-                }
-                else {
-                    pugi::xml_document newDoc;
-                    writer.SaveGeometry(newDoc, model);
-                    //writer.SaveSimulationState(newDoc, model, globState);
-                    writer.SaveXMLToFile(newDoc, outFile);
-                    //SettingsIO::workFile = outFile;
-                }
-                writer.SaveSimulationState(outFile, model, globState);
-            } catch(std::filesystem::filesystem_error& e) {
-                Log::console_error("Warning: Could not create file: %s\n", e.what());
-            }
 
             // if there is a next des limit, handle that
             if(!Settings::desLimit.empty()) {
+
+                // First write an intermediate output file
+                // TODO: Make it toggeable?
+                /*std::stringstream outFile;
+            outFile << SettingsIO::outputPath << "/" <<"desorped_" << model->otfParams.desorptionLimit << "_" <<
+                 std::filesystem::path(SettingsIO::outputFile).filename().string();*/
+                std::string outFile = std::filesystem::path(SettingsIO::outputPath)
+                        .append("desorbed_")
+                        .concat(std::to_string(model->otfParams.desorptionLimit))
+                        .concat("_")
+                        .concat(std::filesystem::path(SettingsIO::outputFile).filename().string()).string();
+
+                try {
+                    FlowIO::WriterXML writer;
+                    Log::console_msg_master(3, " Saving intermediate results: %s\n", outFile.c_str());
+                    if(!SettingsIO::workFile.empty() && std::filesystem::exists(SettingsIO::workFile)) {
+                        try {
+                            std::filesystem::copy_file(SettingsIO::workFile, outFile,
+                                                           std::filesystem::copy_options::overwrite_existing);
+                        } catch (std::filesystem::filesystem_error &e) {
+                            Log::console_error("Could not copy file: %s\n", e.what());
+                        }
+                    }
+                    else {
+                        pugi::xml_document newDoc;
+                        writer.SaveGeometry(newDoc, model);
+                        //writer.SaveSimulationState(newDoc, model, globState);
+                        writer.SaveXMLToFile(newDoc, outFile);
+                        //SettingsIO::workFile = outFile;
+                    }
+                    writer.SaveSimulationState(outFile, model, globState);
+                } catch(std::filesystem::filesystem_error& e) {
+                    Log::console_error("Warning: Could not create file: %s\n", e.what());
+                }
+                // Next choose the next desorption limit and start
+
                 model->otfParams.desorptionLimit = Settings::desLimit.front();
                 Settings::desLimit.pop_front();
                 simManager.ForwardOtfParams(&model->otfParams);
                 endCondition = false;
-                Log::console_msg_master(1, " Handling next des limit %z\n", model->otfParams.desorptionLimit);
+                Log::console_msg_master(1, " Handling next des limit %zu\n", model->otfParams.desorptionLimit);
 
                 try {
                     ProcessSleep(1000);
@@ -197,20 +271,9 @@ int main(int argc, char** argv) {
 
         if(Settings::outputDuration && (uint64_t)(elapsedTime)%Settings::outputDuration==0){ // autosave every x seconds
             if((uint64_t)elapsedTime / Settings::outputDuration <= 1){
-                Log::console_msg_master(1, "\n%-6s %-14s %-20s %-20s %-20s %-20s %-20s %-20s\n",
-                                        "Node#", "Time",
-                                        "#Hits (run)", "#Hits (total)","Hit/sec",
-                                        "#Des (run)", "#Des (total)","Des/sec");
-                Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'-').c_str());
+                printer.PrintHeader();
             }
-            Log::console_msg(1,"%-6d %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-                             MFMPI::world_rank, elapsedTime,
-                             globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-                             (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-                             (elapsedTime),
-                             globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                             (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                             (elapsedTime));
+            printer.Print(elapsedTime, globState);
         }
 
         // Check for potential time end
@@ -230,26 +293,15 @@ int main(int argc, char** argv) {
 #ifdef USE_MPI
     MPI_Barrier(MPI_COMM_WORLD);
 #endif
-    Log::console_msg_master(1, "\n%-6s %-14s %-20s %-20s %-20s %-20s %-20s %-20s\n",
-                            "Node#", "Time",
-                            "#Hits (run)", "#Hits (total)","Hit/sec",
-                            "#Des (run)", "#Des (total)","Des/sec");
-    Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'-').c_str());
+    printer.PrintHeader();
 #ifdef USE_MPI
     MPI_Barrier(MPI_COMM_WORLD);
 #endif
 
-    //TODO: Send output to master node
+    //TODO: Send output to master node for ordered output
     if(elapsedTime > 1e-4) {
         // Global result print --> TODO: ()
-        Log::console_msg(1,"%-6d %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-               MFMPI::world_rank, elapsedTime,
-               globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-               (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-               (elapsedTime),
-               globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                         (elapsedTime));
+        printer.Print(elapsedTime, globState);
     }
 
 #if defined(USE_MPI)
@@ -269,18 +321,11 @@ int main(int argc, char** argv) {
     if(MFMPI::world_rank != 0){
         return 0;
     }
-    Log::console_msg_master(1, "%s\n",std::string(6+14+20+20+20+20+20+20,'=').c_str());
 #endif //USE_MPI
 
     if(elapsedTime > 1e-4) {
-        Log::console_msg(1,"%-6s %-14.2lf %-20zu %-20zu %-20.2lf %-20zu %-20zu %-20.2lf\n",
-                         "x", elapsedTime,
-                         globState.globalHits.globalHits.nbMCHit - oldHitsNb, globState.globalHits.globalHits.nbMCHit,
-                         (double) (globState.globalHits.globalHits.nbMCHit - oldHitsNb) /
-                         (elapsedTime),
-                         globState.globalHits.globalHits.nbDesorbed - oldDesNb, globState.globalHits.globalHits.nbDesorbed,
-                         (double) (globState.globalHits.globalHits.nbDesorbed - oldDesNb) /
-                         (elapsedTime));
+        if(MFMPI::world_size > 1)
+            printer.Print(elapsedTime, globState, true);
     }
 
     if(MFMPI::world_rank == 0){
@@ -306,9 +351,14 @@ int main(int argc, char** argv) {
         }
         else if(!SettingsIO::overwrite){
             // Copy full file description first, in case outputFile is different
-            if(!SettingsIO::workFile.empty())
-                std::filesystem::copy_file(SettingsIO::workFile, fullOutFile,
-                                       std::filesystem::copy_options::overwrite_existing);
+            if(!SettingsIO::workFile.empty() && std::filesystem::exists(SettingsIO::workFile)){
+                try {
+                    std::filesystem::copy_file(SettingsIO::workFile, fullOutFile,
+                                               std::filesystem::copy_options::overwrite_existing);
+                } catch (std::filesystem::filesystem_error &e) {
+                    Log::console_error("Could not copy file to preserve initial file layout: %s\n", e.what());
+                }
+            }
         }
         FlowIO::WriterXML writer(false, true);
         pugi::xml_document newDoc;
